Add 2020 United States Data [READ]
Used data_set_creator.py script in order to generate more cohesive
datasets. I couldn't sift through every single United States fire
because going through the global fire data set would take 1/2 day per
dataset. For the .shp file, I picked a random integer between 0 and the
number of elements in the .shp file (magnitude in the hundreds of
thousands). Then, I looked at the ID of that row and added all of the
other rows with the same ID with higer indices and lower indices (all IDs are grouped together in the
table) ONLY if that fire originated in the United States.

I did the United States check by taking the first longitude/latitude point in
the polygon. Then, I used a Google Cloud API to use Google Maps to map
that longitude/latitude point to a country with some finessing. If the
fire polygon that was randomly selected existed in the United States, I
added it to the dataset. I kept the fireIDs in a set so that a fire
could NOT be added twice.

I stored all of the data through writing a new .shp file. These were
eventually moved to the data repository under "United States 2020
Fires". I did this for 12 datasets so we have a fully generated set of
United States fires during 2020. It is good to note that data for
January 2020 has fires that started in 2019. I think this is because the
"January 2020" is simply when they log things. It doesn't necessarily
mean the fire happened in January 2020.