Add day to day representation of "land units" + more
Firstly, I moved all of the 2020 data to a separate folder, which
matches up with what's on the OneDrive. Note, GitHub only has the 2020
data while OneDrive has 2000 to 2020 data for the United States

Next, I created the framework in the Single_Fire_Analysis.ipynb for how
to represent a multiple day fire at the "land unit" scale (i.e. the
minimum scale our data holds for for the wildfires). This, in effect, is
discretizing our data into a 3D array into a k by x by y matrix. The k's
are the days, the x's are the # of land units for latitude, the y's are
the # of land units for longitude.

The US_data_plotting.py dataset was simply working with how to get this
data for Global Fire simplified down to just US data. I may move all of
the "developmental notebooks" to a different folder because I think it's
causing confusion. The most useful functions that I've created, I'll put
in util.py and push.

I also made small changes to the other notebooks in the development of
this one.