Develop Neural Net for overfitting one fire
As an "experiment", I've implemented a proof of the value of an LSTM in
our model. For a given fire (with lots of data), I've trained a RNN w/
4 ReLU activation hidden layers in order to
guess at what the next days fires will look like given the previous k
number of days of fires.

For k = 1, the RNN gets a value of .9495 accuracy. For k = 3, the RNN
gets a value of .9570. For k = 5, the RNN gets a value of .9735. This
demonstrates that having an LSTM is a beneficial way to approach this
problem for Neural Nets.

However, this method of doing things is kind of sketchy. I'm passing the
previous 12 (m x n) grids of binary fire/no fire and I'm passing the
current grid of (m x n) binary fire/no fire as the output. Since the
fire, I'm training on is only 12 days long, I can only train for 12 - k
fires. This is not nearly enough data, especially since I gave the model
one of the biggest fires I could find. Nevertheless, I think this is
something that I will explore later.

I think this is a good enough "experiment" for now, and hopefully we can
talk about how we can generalize this to training a model on multiple
fires. Right now, it's training on an (m x n) grid for one fire but
different fires have different sized grids, AH.