Can finally generate .npy fire reps quickly
It took a while but fires can now be more quickly generated as numpy
arrays. Before, it took a long, long, long amount of time to iterate
throug the big fires. I optimized a lot of the code for this purpose and
am running it. Will post the outputs in the Google Drive.

The plot_from_numpy_multi_day.ipynb demonstrates plotting from the numpy
arrays.

single_fire.py and single_fire_data_creator.py were both changed for the
optimizations.

Images from the mozambique fire were added that were put in the
midpoint. So was a paper.