Mass Commit
This commit is for a bunch of scripts that I've been working on that
works with (or generates) the new data that we've found.

They are NOT very put together and should NOT be taken as absolute
truths of being written correctly. I just want everyone to have the
opportunity to work with what I've been working with.

Of particular importance are generate-pickle, nn-w-elevation, and
nn-development.

We should discuss all of these in our meeting.